Handle cookbooks with evaluated metadata.
Cookbooks' metadata, as stored in [99% of git repos](https://github.com/opscode-cookbooks), are `.rb` files that can contain references to other files within the repo (typically `README.md` or `VERSION` files).

When Chef server is serving cookbooks, these Ruby-time dependencies are handled by `knife` (i.e. when the cookbook is uploaded to the server), so the server never has to deal with them. We don't have that luxury, though, because we're running from the original source.